# apps/agentService/api/agent_workspace/v1/agent_workspace_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             v3.12.4
// source: agent_workspace/v1/agent_workspace.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAgentWorkspaceAddAgentToWorkspace = "/api.agent_workspace.v1.AgentWorkspace/AddAgentToWorkspace"
const OperationAgentWorkspaceGetWorkspaceAgents = "/api.agent_workspace.v1.AgentWorkspace/GetWorkspaceAgents"
const OperationAgentWorkspaceRemoveAgentFromWorkspace = "/api.agent_workspace.v1.AgentWorkspace/RemoveAgentFromWorkspace"

type AgentWorkspaceHTTPServer interface {
	// AddAgentToWorkspace 添加Agent到工作区
	AddAgentToWorkspace(context.Context, *AddAgentToWorkspaceRequest) (*AddAgentToWorkspaceResponse, error)
	// GetWorkspaceAgents 获取工作区下的Agent列表
	GetWorkspaceAgents(context.Context, *GetWorkspaceAgentsRequest) (*GetWorkspaceAgentsResponse, error)
	// RemoveAgentFromWorkspace 删除工作区中的Agent
	RemoveAgentFromWorkspace(context.Context, *RemoveAgentFromWorkspaceRequest) (*RemoveAgentFromWorkspaceResponse, error)
}

func RegisterAgentWorkspaceHTTPServer(s *http.Server, srv AgentWorkspaceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/agent/workspaces/agents", _AgentWorkspace_GetWorkspaceAgents0_HTTP_Handler(srv))
	r.POST("/v1/agent/workspaces/{agentId}", _AgentWorkspace_AddAgentToWorkspace0_HTTP_Handler(srv))
	r.DELETE("/v1/agent/workspaces/agents/{agentId}", _AgentWorkspace_RemoveAgentFromWorkspace0_HTTP_Handler(srv))
}

func _AgentWorkspace_GetWorkspaceAgents0_HTTP_Handler(srv AgentWorkspaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetWorkspaceAgentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAgentWorkspaceGetWorkspaceAgents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetWorkspaceAgents(ctx, req.(*GetWorkspaceAgentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetWorkspaceAgentsResponse)
		return ctx.Result(200, reply)
	}
}

func _AgentWorkspace_AddAgentToWorkspace0_HTTP_Handler(srv AgentWorkspaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddAgentToWorkspaceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAgentWorkspaceAddAgentToWorkspace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AddAgentToWorkspace(ctx, req.(*AddAgentToWorkspaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddAgentToWorkspaceResponse)
		return ctx.Result(200, reply)
	}
}

func _AgentWorkspace_RemoveAgentFromWorkspace0_HTTP_Handler(srv AgentWorkspaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RemoveAgentFromWorkspaceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAgentWorkspaceRemoveAgentFromWorkspace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RemoveAgentFromWorkspace(ctx, req.(*RemoveAgentFromWorkspaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RemoveAgentFromWorkspaceResponse)
		return ctx.Result(200, reply)
	}
}

type AgentWorkspaceHTTPClient interface {
	AddAgentToWorkspace(ctx context.Context, req *AddAgentToWorkspaceRequest, opts ...http.CallOption) (rsp *AddAgentToWorkspaceResponse, err error)
	GetWorkspaceAgents(ctx context.Context, req *GetWorkspaceAgentsRequest, opts ...http.CallOption) (rsp *GetWorkspaceAgentsResponse, err error)
	RemoveAgentFromWorkspace(ctx context.Context, req *RemoveAgentFromWorkspaceRequest, opts ...http.CallOption) (rsp *RemoveAgentFromWorkspaceResponse, err error)
}

type AgentWorkspaceHTTPClientImpl struct {
	cc *http.Client
}

func NewAgentWorkspaceHTTPClient(client *http.Client) AgentWorkspaceHTTPClient {
	return &AgentWorkspaceHTTPClientImpl{client}
}

func (c *AgentWorkspaceHTTPClientImpl) AddAgentToWorkspace(ctx context.Context, in *AddAgentToWorkspaceRequest, opts ...http.CallOption) (*AddAgentToWorkspaceResponse, error) {
	var out AddAgentToWorkspaceResponse
	pattern := "/v1/agent/workspaces/{agentId}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAgentWorkspaceAddAgentToWorkspace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AgentWorkspaceHTTPClientImpl) GetWorkspaceAgents(ctx context.Context, in *GetWorkspaceAgentsRequest, opts ...http.CallOption) (*GetWorkspaceAgentsResponse, error) {
	var out GetWorkspaceAgentsResponse
	pattern := "/v1/agent/workspaces/agents"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAgentWorkspaceGetWorkspaceAgents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *AgentWorkspaceHTTPClientImpl) RemoveAgentFromWorkspace(ctx context.Context, in *RemoveAgentFromWorkspaceRequest, opts ...http.CallOption) (*RemoveAgentFromWorkspaceResponse, error) {
	var out RemoveAgentFromWorkspaceResponse
	pattern := "/v1/agent/workspaces/agents/{agentId}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAgentWorkspaceRemoveAgentFromWorkspace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}